Quick fix for broken use of M_PI on MSVC.
The symbol M_PI is not defined by default in MSVC.  I've added a #define
that will force it to be defined.  However, this isn't really a solid
fix.  If <math.h> happens to be included before <cusp/complex.h> then
M_PI will be undefined again.  It would be better to find a more
solid and portable fix for this.